+++ Sources/HomomorphicEncryption/Bfv/Bfv+Decrypt.swift
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 import Foundation
+import ModularArithmetic
 
 extension Bfv {
     @inlinable
