+++ Sources/HomomorphicEncryption/Bfv/Bfv+Encrypt.swift
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import ModularArithmetic
+
 extension Bfv {
     @usableFromInline
     enum PlaintextTranslateOp {
