+++ Sources/HomomorphicEncryption/Bfv/Bfv+Keys.swift
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import ModularArithmetic
+
 extension Bfv {
     @inlinable
     // swiftlint:disable:next missing_docs attributes
