+++ Sources/HomomorphicEncryption/CoefficientPacking.swift
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import ModularArithmetic
+
 /// Utilities to transcode between bytes and scalar coefficients.
 public enum CoefficientPacking {}
