+++ Sources/HomomorphicEncryption/Encoding.swift
@@ -15,6 +15,8 @@
 // The encode/decode functions in context are not supposed to be called directly. Instead, one should call the
 // corresponding encode/decode functions in specific Scheme instead.
 
+import ModularArithmetic
+
 extension Context {
     /// Encodes `values` in the given format.
     ///
