+++ Sources/HomomorphicEncryption/Plaintext.swift
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import ModularArithmetic
+
 /// Plaintext type.
 public struct Plaintext<Scheme: HeScheme, Format: PolyFormat>: Equatable, Sendable {
     public typealias SignedScalar = Scheme.SignedScalar
