+++ Sources/HomomorphicEncryption/PolyRq/Galois.swift
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import ModularArithmetic
+
 @usableFromInline
 struct GaloisCoeffIterator: IteratorProtocol {
     @usableFromInline typealias Element = (Bool, Int)
