+++ Sources/HomomorphicEncryption/PolyRq/PolyContext.swift
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import ModularArithmetic
+
 /// Polynomial context that holds all the pre-computed values for doing efficient calculations on ``PolyRq``
 /// polynomials.
 public final class PolyContext<T: ScalarType>: Sendable {
