+++ Sources/HomomorphicEncryption/PolyRq/PolyRq+Ntt.swift
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import ModularArithmetic
+
 extension ScalarType {
     /// Returns whether or not a prime value is a valid NTT modulus.
     ///
