+++ Sources/HomomorphicEncryption/PolyRq/PolyRq.swift
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import ModularArithmetic
+
 /// Represents a polynomial in `R_q = Z_q[X] / (X^N + 1)` for `N` a power of
 /// two and `q` a (possibly) multi-word integer.
 ///
