+++ Sources/HomomorphicEncryption/RnsTool.swift
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import ModularArithmetic
+
 @usableFromInline
 package struct RnsTool<T: ScalarType>: Sendable {
     /// `Q = q_0, ..., q_{L-1}`.
