+++ Sources/PrivateInformationRetrieval/MulPir.swift
@@ -14,6 +14,7 @@
 
 import Foundation
 import HomomorphicEncryption
+import ModularArithmetic
 import Numerics
 
 /// PIR using ciphertext-ciphertext multiplication.
